requirements.txt
streamlit>=1.36.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.20.0
matplotlib>=3.8.0
prophet>=1.1.5
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.30.0
oauth2client>=4.1.3
scikit-learn>=1.4.0
statsmodels>=0.14.0
sqlalchemy>=2.0.0
openai>=1.40.0
python-dotenv>=1.0.1
bcrypt>=4.1.3
PyYAML>=6.0.1
openpyxl>=3.1.0
